+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - (docs) added `demo-readme` example to `demo-notebooks` section,
 - (docs) demo example in documentation updated,
 - (docs) link to the documentation page added to README,
-
+- (enhancement) item-sessions map may be derived from session-items map,
+- (feature) data can be read from the flat record structure,
+- (feature) data can be parsed from dataframes
 
 ## Version 1.1.1 (2023-07-08)